move some of api to common utils
